节点配置
